Move domain verification into VercelProject
